Adding the course instructor tables
I copied the tables to a new folder called ImportedResults. The idea is that this folder can contain results that we created on IPEA or elsewhere. In the long run, this folder probably shouldn't exist because we want to be more systematic.

Also committing the version of the paper with filepaths fixed up through the course instructor tables.