Create add_ids_as_vertex_property.py
Create a new function for adding vertex IDs as internal vertex properties. This is useful when the vertices are meaningful, for example gammas, and we want to preserve a mapping between the original IDs in the edge list we used to create the graph and the internal graphtool vertex IDs that are continuously numbered, starting at 0.