Update the _recode variables to go from1 to N+1 instead of 0 to N
These variables were coded from 0 to N, so 0 represented an ordinary category. However, iota and gamma are both coded from 1 to N+1 with -1 representing missing and gamma=0 corresponding to non-employment (there is no iota=0 because non-employment doesn't make sense as a state for classifying a worker's type in our context; only for classifying the job they're currently matched with).

I think the solution, which I implemented here, is to have all of the _recode variables range from 1 to N+1 rather than 0 to N. I'll have to think about whether or not this causes any problems elsewhere, but the fact that iota and gamma are already coded this way implies that it should be fine.